+++ crates/mimic/src/core/traits.rs
@@ -91,20 +91,23 @@ pub trait EntityKind: TypeKind + EntitySearch + EntitySort {
     const STORE: &'static str;
     const INDEXES: &'static [EntityIndex];
 
+    // values
+    fn values(&self) -> Values;
+
     // primary_key
     // returns the current primary key (a slice of IndexValues)
     fn primary_key(&self) -> Self::PrimaryKey;
 
     // entity_key
-    // returns the current runtime entity key, ie [V::Text("00AX5"), V::Nat8(1)]
+    // the runtime EntityKey is the PrimaryKey as a vector
     fn entity_key(&self) -> ::mimic::core::types::EntityKey {
         ::mimic::core::types::EntityKey(self.primary_key().as_ref().to_vec())
     }
 
     // data_key
-    // builds the data key using the current entity key
+    // builds the data key using the current primary key
     fn data_key(&self) -> DataKey {
-        Self::build_data_key(&self.entity_key())
+        Self::build_data_key(self.primary_key().as_ref())
     }
 
     // build_data_key
@@ -133,9 +136,6 @@ pub trait EntityKind: TypeKind + EntitySearch + EntitySort {
             None
         }
     }
-
-    // values
-    fn values(&self) -> Values;
 }
 
 ///