+++ crates/mimic_design/src/node/newtype.rs
@@ -2,7 +2,7 @@ use crate::{
     helper::quote_option,
     imp::{self, Imp},
     node::{Arg, Def, Item, MacroNode, Node, Trait, TraitNode, TraitTokens, Type},
-    schema::{BPrimitiveType, Schemable},
+    schema::{BPrimitive, Schemable},
     traits::Traits,
 };
 use darling::FromMeta;
@@ -18,7 +18,7 @@ pub struct Newtype {
     #[darling(default, skip)]
     pub def: Def,
 
-    pub primitive: BPrimitiveType,
+    pub primitive: BPrimitive,
     pub item: Item,
 
     #[darling(default)]